Real-Time Image Captioning and Voice Narration using BLIP and gTTS/requirements.txt
torch>=1.12.0
transformers>=4.30.0
opencv-python
Pillow
gTTS
playsound